Added signup/login + some firebase installation
I installed the start of the firebase nodes. I did it in the frontend folder because it needed like the package.json, and it installs its nodes inside the same one react does, so thats why I did that. We can probably move it if we need. In addition we still need to finish firebase installation so I can run the emulator. Also the login page needs to be able to redirect to the signup page, and then the rest should work.