DisplayName and other fixes
Fixed display name so you enter it when you sign up, and it sticks. Also added secure password functionality. The only problem is with the password reset link, the only amount of security it requires is 6 characters, so since it's like an external firebase link, we can only fix it if we somehow either check with the database some way or if we figure out how to send a custom link that we host, but I'm not too sure how to do that right now.